Extract scripts from npm package
This moves things around in a more common idiomatic folder structure / naming convention. This extracts the scripts that are only used to test the library from the source of the library itself.

/src/scripts -> /scripts

This is no longer part of the build process. Therefore we use babel-node to run the files as needed. The exception is test-runner that is just ran in plain node so that it works with the coverage reports.

The lib folder now includes only and all the library files required to use the package and no longer test runners.

Because of that I moved all the dependencies that were related only to our test scripts and website build to devDependencies. That way we minimize the required dependencies to use Prepack downstream as an npm package.

Then I configured the package.json to only list lib/ and bin/ folders as part of the package. This lets `npm publish` ignore all other files so that people installing the package don't need to include all our testing code.

I added files to bin/ and package.json which is a convention to make this package installable and usable as a command-line tool.

Renamed run_util -> prepack-cli and repl -> repl-cli to separate these from library variants.

The coverage reports now get created in the root instead of lib folder so that they don't get included in the build.

I bumped the version number to one minor above what is already on npm since we can't publish earlier than that.