Move $ContextStack to $ErrorData and append location data to syntax errors (#461)
I believe the $ErrorData slot is meant to be an internal host provided
object that keeps things like the context stack on it. So we don't need
to come up with another slot to store it. We might want to minimize
internal fields to create monomorphic data structures as a perf
optimization.

I also added another field to $ErrorData that stores the source location
of syntax errors when the parser causes an error.

This is used inside hosting environments such as Node to print a nicer
error message with the location of the syntax error. This is different
from the contextStack since this is not actually part of the execution
stack.

This can be used to print messages with an error showing the location of
the syntax error.

```
foo.js:3
x ++ y
     ^
SyntaxError: Unexpected identifier
```