Use public APIs in tests
This gives us some better end-to-end coverage and makes it easier move
internals around.

This surfaced a few bugs/issues in the API.

I left the thing that detects introspection errors but this surfaced that
we don't have a way to pass custom error loggers through the public API.
I didn't want to expose the whole Realm and Value types externally. But
it might be worth having a string based way to get logs.