Use an existing property name to test throwing in global getters (#566)
This works around in a bug in Node.js 7.x.x where added properties need to
be copied over onto the original "sandbox" created by the VM module in the
test-runner. That is done in native code which is not resilient to this
getter throwing which causes a crash.

By using an existing property name, that path is skipped.