Update Circle CI Node Version to 7.9.0 (#576)
The purpose of this is to be able to land #397 with some test coverage.
In its current form it is too specific in its implementation and only
works on a specific version of Node. In the future we'll want to make that
more abstract and general so that it can work on ranges and other major
versions.