Fixing regression around delaying of "referentialization"
Summary:
We recently added a feature that delays "referentialization" (#643),
which is the allocation of mutable slots for captured variables.
However, this caused a regression when the same captured variable
gets involved in multiple functions, with the delaying logic kicking
in for some functions but not others.

This change ensures that referentialization is always delayed in a
uniform way.

This doesn't always produce optimal code, but that's for another pull request.

Added regression test.
Closes https://github.com/facebook/prepack/pull/741

Differential Revision: D5282069

Pulled By: NTillmann

fbshipit-source-id: 97b47c0da41c956b2a44376942877672ae6cbe9e