Early termination for fatal errors
Summary:
If an error is reported to the host application and it returns a value that indicates that the error is fatal, the evaluator now throws a FatalError exception rather than causing an IntrospectionError.

This terminates things quickly and cleanly but uncovered a bug in the way effect tracking is cleaned up in the face of exceptions that are not throw completions. To fix this, the code that pops an evaluation context off the stack now checks if there are any dangling effects in the context and, if there are, it folds them into the effects of the outer context.

The effects then propagate to the most closely nested evaluateForEffects call, where they are
rolled back from the global state and incorporated into the effects returned from the call.

When they propagate all the way to a test runner, as happens when there is an IntrospectionError in global code, the state is never rolled back and thus the error can be logged in the state that applied when it was created.

DEPRECATED API
The InitializationError constructor in prepack-standalone.js is going to go away in a future release. Please use FatalError instead. For now, InitializationError.prototype is on the prototype chain of FatalError.prototype, so instances of FatalError will still be instances of InitializationError, so this should not be a breaking change in the next release.
Closes https://github.com/facebook/prepack/pull/740

Differential Revision: D5286107

Pulled By: hermanventer

fbshipit-source-id: 05c7f9197acaa0ba922d136f122968b2f41a4b82