Delay serialization of any value that is only referenced from residual functions
Summary:
This will address #496.

Done so far: Keep track of the scopes (in terms of functions) in which values are needed.
Still to be done: Use that information in the serializer to move initialization of values only used by single residual function into a separate block that run once inside of the residual function.
Closes https://github.com/facebook/prepack/pull/699

Differential Revision: D5295087

Pulled By: hermanventer

fbshipit-source-id: dddb7ed2ad9db6307a3851a2b7598c04d5ce8505