Error message for abstract with object
Summary:
If a with object is abstract always give an error message.

This is a bit stricter than before, because partial non-simple objects will also be faulted, but it is a better user experience because the error points to the with object expression and talks directly about the problems with with, rather than possibly failing obscurely when trying to access a name that is not defined by the with object.

This addresses issue: #743
Closes https://github.com/facebook/prepack/pull/744

Differential Revision: D5297182

Pulled By: hermanventer

fbshipit-source-id: 8f1427b47ae6d636b5a08290a158fad8aad610d0