Error messages for object expressions
Summary:
Generalize object expressions to allow abstract values as property names, since it is already possible to use abstract property names via `ob[absPropName] = val`.

In the case where the abstract value is not known to be a string value, give an error message since this case is not supported by either syntax (and is quite a bit more complicated to support).

[PP0011 property key value is unknown](https://github.com/facebook/prepack/wiki/PP0011)
Closes https://github.com/facebook/prepack/pull/751

Differential Revision: D5322785

Pulled By: hermanventer

fbshipit-source-id: a929f730d01deca1ae1a3299475c3c0bc6d5cde0