Check that abstract reference base objects are coercible to objects
Summary:
The base value of a Reference is an abstract value that is not known to be an object or one of a set of objects, then the effects of an assignment to the reference is unknown to the extent that Prepack cannot continue. It also had the effect that an invariant was violated, leading to issue #750.

We now give a error message that is contextualized to the member expression (if any) that created  the Reference.

[PP0012 reference base value is unknown](https://github.com/facebook/prepack/wiki/PP0012)
Closes https://github.com/facebook/prepack/pull/756

Differential Revision: D5331077

Pulled By: hermanventer

fbshipit-source-id: 794e5e15eff481fef5cf2f1fe0b90f7dcde325e6