Track property changes due to Object.defineProperty
Summary:
Tracking did not work for this because the convoluted logic for this method mutated the old property descriptor before registering it needing to get tracked. (The tracking logic requires property updates to always use completely new descriptors.)

Also enabled joining for get and set properties of descriptors, so that the test case does not fail. This is not something we've seen in the wild so far.

Issue: #765
Closes https://github.com/facebook/prepack/pull/766

Differential Revision: D5345148

Pulled By: hermanventer

fbshipit-source-id: 71fce1d541370ffb5e33aa9b615feece99a4fe94