Updated Sandcastle Tests
Summary: Sandcastle Tests now prepack a checked in version of instagram on every single diff. Of note is that to run the test-internal locally you need to put the instagram bundles in a directory called facebook/test as this is the cleanest way to run these tests internally. The buck targets are updated and still exist but are not entirely necessary right now. If people outside of prepack want to build it and use it, however, they could link up with the buck target.

Reviewed By: cblappert

Differential Revision: D5335451

fbshipit-source-id: 102cb228e0b422b399605132de96b881cd4233e7