Doing the right thing for "global." in __abstract names.
Summary:
The `global` identifier isn't always defined in JavaScript. In Prepack, it is meant to refer to the global `this` object. Now, when referencing `global.` in the given expression for an abstract value, the appropriate transformation is done.

Besides enabling this for the `__abstract` function, this pull request also updates other existing references such as `Date.now()` to explicitly refer to `global.Date.now()`.
Closes https://github.com/facebook/prepack/pull/773

Differential Revision: D5357294

Pulled By: NTillmann

fbshipit-source-id: 6c8a7525ebd45058e0f81c63c9727e1c006b5e66