Support throw completions that escape functions bodies
Summary:
The code for throwing an introspection error if an conditional throw completion occurs, kicks in too late for one of the invariants. This invariant now fails for one of our scenarios because of the inclusion of the invariant module.

Fixing this so that the invariant is preserved turns out to be tricky, so I've had to implement more of use cases for PossiblyNormalCompletions. Calls that return the latter (because of a conditional, uncaught throw in the body), now no longer terminate Prepack immediately, but execution proceeds until the end of the statement containing the call. When the completion resulting from that statement is inspected and found to be a PossiblyNormalCompletion, an IntrospectionError is thrown.

When running the partial evaluator, things carry on.
Closes https://github.com/facebook/prepack/pull/783

Differential Revision: D5377111

Pulled By: hermanventer

fbshipit-source-id: 094439acfeb49f2e96dcb5f184b847228b9a9386