Make it clear when a node can result in a Reference
Summary:
Currently the evaluate implementation of every node claims that it could result in a Reference. In fact, only Identifier, Member and SuperProperty can result in a Reference. To make this more obvious, this request removes Reference from the type signature of every evaluate implementation where the result can never be a Reference. It also removes Value from the type signature of Identifier, Member and SuperProperty.
Closes https://github.com/facebook/prepack/pull/797

Differential Revision: D5403889

Pulled By: hermanventer

fbshipit-source-id: 0316f34512d7f92995e475b33928a983fb13f14c