Always throw FatalErrors or use invariants
Summary:
This cleans up the current uses of "throw Error('some string');" turning some into invariants (when the throw could only happen if there is a programming bug) and others into "throw new FatalError" (when the throw could happen because an input program contains something that is not yet supported).

In the fullness of time the second category of errors need to be paired up with handleError calls and some of them will go away as more features are implemented.
Closes https://github.com/facebook/prepack/pull/813

Differential Revision: D5430100

Pulled By: hermanventer

fbshipit-source-id: e8067c0e5d850e58b2706113d97447f06c258bb6