Specify additional initialization functions to prepack
Summary:
Provide a new option to specify a number of additional functions to prepack (in addition to the global code "function", that is). These functions need to exist in the global scope, have no parameters, be well behaved (never throw an exception) and should not depend on each other (the order in which they get called should not matter).

At the moment there are checks only for existence, good behavior and no write-write conflicts. Also, the functions are not actually prepacked yet.
Closes https://github.com/facebook/prepack/pull/846

Differential Revision: D5514569

Pulled By: hermanventer

fbshipit-source-id: e2e343d9eed4772fc11810d91293facdafca8fcb