Fixing a recent regression
In the refactoring #815, a subtle semantic change caused an invariant
around the `_waitingForEmptyStack` field of the Emitter to be broken.
While investigating the purpose of `_waitingForEmptyStack`, I concluded
that this entire concept was just there in order to work around
the design decision that delay-reasons and values-declared-by-generator-entries
were based on babel identifiers, and not the `Value`s themselves.

So I changed that design decision, made it all `Value` based.
Now we no longer need the dubious `true` delay reason, which used to
indiscriminately wait until *all* actively emitted values (those on the stack)
were available. Instead, as it's easy to name the `Value` that we actually
need to wait for, that's what we do now.

Without the need for the `true` delay reason, the `_waitingForEmptyStack`
concept is no longer needed.

While at it, I also did some trivial renamings.