Fix a recent regression
Summary:
In the refactoring #815, a subtle semantic change caused an invariant
around the `_waitingForEmptyStack` field of the Emitter to be broken.
While investigating the purpose of `_waitingForEmptyStack`, I concluded
that this entire concept was just there in order to work around
the design decision that delay-reasons and values-declared-by-generator-entries
were based on babel identifiers, and not the `Value`s themselves.

So I changed that design decision, made it all `Value` based.
Now we no longer need the dubious `true` delay reason, which used to
indiscriminately wait until *all* actively emitted values (those on the stack)
were available. Instead, as it's easy to name the `Value` that we actually
need to wait for, that's what we do now.

Overall, this will actually slightly change the generated code, hopefully for the better (in terms of local variable lifetimes), as we will now emit certain operations slightly more eagerly, closer to the declaration.

Without the need for the `true` delay reason, the `_waitingForEmptyStack`
concept is no longer needed.

While at it, I also did some trivial renamings, and speeding up serialization by
- avoiding potentially slow `this._stack.indexOf(val)` recursion checks, instead maintaining a set.
- prefer naming a compound value itself as the delay reason, and not one of its components

Adding regression test.
Closes https://github.com/facebook/prepack/pull/853

Differential Revision: D5530131

Pulled By: NTillmann

fbshipit-source-id: 348e2ff83151bc2d5bbbd0718f6960ff5d2c3a1d