Does deduplication of pure builtin calls in generator
Summary:
`Date.now()` and `Math.random()` calls emitted into the generator now no longer get serialized if their results are not used. In principle this is safe for any pure definitions of AbstractValues emitted to the generator but that is too hard to determine for most cases.

Adds a test case verifying that extra `Date.now()` calls are eliminated.

See #543 for more detail.
Closes https://github.com/facebook/prepack/pull/822

Differential Revision: D5537430

Pulled By: cblappert

fbshipit-source-id: 4f3e12eee65fdbea7504d73399eddde5a1e93703