Add JSXElement extension to IsFunctionDefinition
I don't want to add the full implementation upstream right now since it
varies by framework and changes with the framework life-cycle.

However, since this is a non-forkable point I'll need to add this extension
upstream. I need this upstream since otherwise I'll need to fork and use
my fork as the npm dependency for experiments.

I don't really have a good idea how to test this since we don't have a test
harness for the interpreter alone and I can't run the same code in Node so
the serializer tests won't work.