Move away from object templates, part 1
Summary:
This is the first in many steps needed to get rid of object templates for abstract objects.

When we get there, AbstractObjectValue will represent either a set of concrete objects, or a completely unknown object. Since any operation performed on the latter will have unknown effects on the heap, most operations will become errors. I.e. completely unknown objects can be assigned to other properties and tested for membership of Object and pretty much nothing else.

Most of the current uses of AbstractObjectValue + template will have to become uses of concrete objects that have been marked as partial and/or simple.

Currently, not all such uses can be simply transformed in this way, because there are  lot of places where checks depend on an object being abstract. This is actually a pretty bad bug, since partial concrete objects are just as problematic. This will be addressed in a series of subsequent pull requests.

The relevant is issue is #882.
Closes https://github.com/facebook/prepack/pull/899

Differential Revision: D5676531

Pulled By: hermanventer

fbshipit-source-id: 816400dc124e0163943c3dfe732c07237aefaef4