Move away from object templates, part 3
Summary:
The getElements function in the values domain requires its argument to not be Top. For AbstractObjectValues it is currently the case that the values domain is never Top. This will soon change, so I'm reviewing all of the places where getElements are called.

These two places are clearly wrong, so I'm just removing them.
Closes https://github.com/facebook/prepack/pull/902

Differential Revision: D5695493

Pulled By: hermanventer

fbshipit-source-id: 172185e1f43216f1d71b24b9c86a543ab2be29e5