Fix serializer wait for generator issue
Summary:
Fix #892
After long investigation, I found the root cause is not that residual function is generated in global scope but the closure binding being emitted in global scope needs to wait for it's dependencies in sub-generator to be inserted in final code stream.
Our emitter can only check if the dependencies are declared(serialized) or not, but it has no capability to know if the sub-generator body that the dependencies emitted in has been inserted into global generator or not. This caused all sort of ordering issue.(All the tests added in this PR failed in master).

To fix this issue, I did the following changes:
1.  Modify Emitter._activeBodies to be a generator stack so that we know the ordering of different generator.
2. Add GeneratorBody as a second reason for waiting dependency if the dependency value is emitting in an active generator lower in generator stack.
3. Distinguish two modes of emitter: emitting in random generator/function or emitting in current active generator(top in the stack).
Closes https://github.com/facebook/prepack/pull/917

Differential Revision: D5708600

Pulled By: yinghuitan

fbshipit-source-id: 2fcbace20177f0404bfc56df71c091668a81b757