Make it easier to combine abstract values via operators. Part 1.
Summary:
In this part, the code for creating an abstract value that results from a binary operator is centralized into AbstractValue.createFromBinaryOp. As part of this, TypeDomain and ValuesDomain now have transfer functions for computing the domain values that result from binary operations.

If this pattern is acceptable, I'll extend it to the other operators.
Closes https://github.com/facebook/prepack/pull/937

Differential Revision: D5734009

Pulled By: hermanventer

fbshipit-source-id: 0e775116b86df5de5da805de241436ecdc575605