Fix property lookup on simple partial objects
Summary:
OrdinaryGetOwnProperty used to just return undefined when asked for the value of a non existent property of a simple partial object. That meant that getting the value "__makeSimple(__makePartial({})).gotcha" deterministically evaluated to undefined at Prepack time, which is all sorts of wrong, as well as inconsistent with what happened in the case of "__makeSimple(__abstract({}, "xxx").gotcha".

Things should now be consistent and the property should evaluate to an abstract value at Prepack time.

This still leaves a whole lot of other broken behaviors when doing Reflection on simple partial objects. But that is already broken, something of a corner case and a whole lot harder to check for, so it will have to wait.
Closes https://github.com/facebook/prepack/pull/925

Differential Revision: D5740483

Pulled By: hermanventer

fbshipit-source-id: c3405fbcce538ec9d9c56392de022c63ccae80d4