Tweak domains
Summary:
TypesDomain.unaryOp was too pessimistic for operands of unknown type. In fact, the type of the operand is immaterial, so just stop checking it.

Also tweak the ValuesDomain to perform operations with the Realm in read only mode. This is less conservative than unaryOp is now and fixes the hole in binaryOp while still allowing BinaryExpression to use computeBinary in its effect-full evaluation.
Closes https://github.com/facebook/prepack/pull/940

Differential Revision: D5746973

Pulled By: hermanventer

fbshipit-source-id: b22129b849865b7cc78198873781abce1f3ada9a