Add testcase to cover FunctionValue undelay closure binding ordering
Summary:
I did some change in _serializeValueFunction() which did not wait for closure binding, then removed firstUsgae checking. This is clearly wrong but no tests failed which means we are missing some tests for covering this scenario.
I came up with this testcase to cover the test coverage gap.
Closes https://github.com/facebook/prepack/pull/948

Differential Revision: D5789833

Pulled By: yinghuitan

fbshipit-source-id: f4b344bc60df592c58b0fe1b1a92771f94f26513