Do some CSE in the serializer
Summary:
Added Value.getHash() and AbstractValue.equals so that we can quickly find abstract values that are structurally equivalent and thus coalesce them into a single expressions. (This is safe because abstract values must correspond to expressions that depend only on the final state and have no side effects.)

Added logic to the serializer code to find and substitute equivalent expressions.

Issues #825 and #597.
Closes https://github.com/facebook/prepack/pull/880

Differential Revision: D5794540

Pulled By: hermanventer

fbshipit-source-id: 12597ca0040dcae3feff8dc1a2228ee866dde279