Move away from object templates, part 5
Summary:
Modify JSON.stringify to work on partial objects rather than with the template object of an abstract object. Modify JSON.parse to produce a partial object rather than an abstract object when parsing the output of stringify called on a partial object.

The relevant test cases for this need partial objects with intrinsic names, so also add a new model function "__makeIntrinsic" that adds an intrinsic name to a given object and fixes up all of the properties so that they are referenced via expressions rooted in the intrinsic name.

With this in place, remove all of the remaining uses of __abstract applied to a model object.

Closes https://github.com/facebook/prepack/pull/913

Differential Revision: D5700865

Pulled By: hermanventer

fbshipit-source-id: c28ac5d9fc2545405b7745f8cc35373116e172ad