Provide a way to join Undefined with an AbstractValue
Summary:
The feature (#931) we need is to have a way to specify that properties obtained from the environment may be undefined. The proposal here is to introduce "global.__optional" which returns an abstract value that can be undefined (and thus does not cause comparisons to undefined to constant fold to false). For example "__optional(__abstract("function", "foo"))".

To make this useful, the result of __optional needs to be refined inside the true branch of a check that foo is not undefined. See Optional2.js for a test case that will not work without refinement.

Unfortunately, refinement is currently extremely simplistic, so the result of __optional must be expressed as a join between undefined and the argument of __optional, leading to an AbstractValue with kind === "conditional" and with a condition that has the same hash code and structure as "foo !== undefined".

This leads to a rather disconcerting self reference, which uncovered a weakness in the code for extracting error information from abstract values. (Also fixed in this request.)

I don't see a better way to things in the near future, so I'd like to go with this until it (hopefully) becomes easy to get back to a situation where every abstract value is a tree.
Closes https://github.com/facebook/prepack/pull/959

Differential Revision: D5818177

Pulled By: hermanventer

fbshipit-source-id: e7be1726793c3538513e9e82879b155a8f95f1b2