Retain equality tests for intrinsic objects
Summary:
Intrinsic objects are hard to model as joins. It turns out that we don't have to. The main thing to do is to not optimize away null/undefined checks for intrinsic objects.

That still leaves us without a way to tell Prepack that an intrinsic value may be null or undefined, but currently there is no scenario where this information will be of use anyway.

This is a very small point change that unblocks the React compiler without any effect on the serializer.

Issue: #966
Closes https://github.com/facebook/prepack/pull/973

Differential Revision: D5859699

Pulled By: hermanventer

fbshipit-source-id: 7c4ee35b079079500ab870904e27db27c5f74524