Enable change tracking for $MapData arrays
Summary:
Map values that are updated in the context of unknown conditions now have change tracking for the elements of the $MapData property and the array values in these properties can now be joined.

Some rudimentary support for Flow typing property bindings with values of type Array is also provided. It resorts to using any for the element type because Flow seems unable to deal with the union of two array types.

Most uses of the value property of a property binding object come from properties that can never have (native) arrays as values, so many invariants to the effect are needed.

Partially addressed #573
Closes https://github.com/facebook/prepack/pull/1040

Differential Revision: D5966066

Pulled By: hermanventer

fbshipit-source-id: 8ceea960f780d63d185d24202e316d310a2e4ea5