Propagate conditional throw completions across calls
Summary:
All of this complexity is because of the need to separate complex completions (containing conditional throw completions) that arrive at the join point at the end of a function body into a separate return completion with a joined value, along with a now possibly normal completion that collects all of the throw completions.

Since the call expression needs to return a value, not a completion, the return completion is thrown away and its value is returned, while the possibly normal completion is stashed in the current execution environment. From there the latter is eventually copied to the caller environment and it is joined with any further completions until the next join point arrives.

If this next join point is a catch, the possibly normal completion will be consumed and the field in the current execution environment will be cleared (unless of course, the catch clause has a conditional abrupt completion).
Closes https://github.com/facebook/prepack/pull/1062

Differential Revision: D6003059

Pulled By: hermanventer

fbshipit-source-id: 91780ae6a051039608c19eb049cdab02652bb73e