Support running test in inferior JavaScript repl for tests.
Define two new flags for test-runner.js:
 --repl <path> : If present the, the prepack sources of test are evaluated
using by piping to the process instead of in a new node context.
--es5 : run babel on tests, used if the repl does not support es6.

To further deal with es6 a new test header flag is introduced, es6. If a test is
not relevant in a non es6 world, the test is skipped if the es5 flag is set.
This flag has been added to some tests.

Running subprocess for each test is obviously more expensive so the default
behavior is unchanged.