Let + deal with implicit conversions that result in abstract values
Summary:
The + operator already checks if its operands are abstract and generates a suitable abstract value as the result in that case. In the case of concrete operands that become abstract after conversion, however, it just gave up.

This pull request fixes this by introducing a variation of ToPrimitive that can result in an abstract value and then using that in the place of ToPrimitive in computeBinary and dealing with the case where the result is abstract.
Closes https://github.com/facebook/prepack/pull/1108

Differential Revision: D6182523

Pulled By: hermanventer

fbshipit-source-id: 9e550f6f7fab8ee5e8974228a6b0451fb549893e