Store Flow class component parameter annotations
Summary:
For the React folding optimizations to work on ES2015 class components, we need to get the Flow type annotations for the `Component`. For example:

```jsx
type Props = { ... };
type State = { ... };

class MyComponent extends React.Component<Props, State> {
  render() { ... }
}
```

The type annotations are used in the reconciliation process as hints for the shape of abstract objects.
Closes https://github.com/facebook/prepack/pull/1128

Differential Revision: D6238366

Pulled By: trueadm

fbshipit-source-id: 776135e3b5d479d21a6b2d32fdacae3d63852f16