Taint objects passed to impure functions
Introducing a new concept of tainting objects and bindings as unknown
after they've been passed to an impure function.

After they have been tainted all mutations are tracked through a generator.

The object itself is serialized as empty. Properties are added to it right
before the call to the first impure function.